Move files from `.~/components/paid-ads/campaign-preview/images` to `.~/images/campaign-preview`.
