+++ js/src/components/paid-ads/campaign-preview/campaign-preview.js
@@ -25,8 +25,8 @@ import MockupSearch from './mockup-search';
 import MockupGmail from './mockup-gmail';
 import MockupDisplay from './mockup-display';
 import MockupMap from './mockup-map';
-import productSampleImageURL from './images/product-sample-image.jpg';
-import shopSampleLogoURL from './images/shop-sample-logo.png';
+import productSampleImageURL from '.~/images/campaign-preview/product-sample-image.jpg';
+import shopSampleLogoURL from '.~/images/campaign-preview/shop-sample-logo.png';
 import { glaData } from '.~/constants';
 import './campaign-preview.scss';
