+++ js/src/images/campaign-preview/map-background.png
